requirements-test.txt
# Testing dependencies for ACEest Fitness application

# Core testing framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0

# Code coverage
coverage[toml]==7.3.2

# Testing utilities
faker==20.1.0
factory-boy==3.3.0

# HTTP testing
requests==2.31.0
requests-mock==1.11.0

# Application dependencies (required for tests)
-r requirements.txt
